+++ packages/web/package.json
@@ -28,6 +28,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.2.0",
+    "zod": "^3.22.2",
     "zustand": "^3.6.9"
   },
   "devDependencies": {
